+++ lib/udap_security_test_kit/endpoints/mock_udap_server/registration_endpoint.rb
@@ -1,12 +1,14 @@
 # frozen_string_literal: true
 
-require_relative '../../urls'
 require_relative '../../tags'
 require_relative '../mock_udap_server'
+require_relative 'udap_response_creation'
 
 module UDAPSecurityTestKit
   module MockUDAPServer
     class RegistrationEndpoint < Inferno::DSL::SuiteEndpoint
+      include UDAPResponseCreation
+
       def test_run_identifier
         MockUDAPServer.client_uri_to_client_id(
           client_uri_from_registration_payload(MockUDAPServer.parsed_io_body(request))
@@ -14,22 +16,7 @@ module UDAPSecurityTestKit
       end
 
       def make_response
-        parsed_body = MockUDAPServer.parsed_io_body(request)
-        client_id = MockUDAPServer.client_uri_to_client_id(client_uri_from_registration_payload(parsed_body))
-        ss_jwt = request_software_statement_jwt(parsed_body)
-
-        response_body = {
-          client_id:,
-          software_statement: ss_jwt
-        }
-        response_body.merge!(MockUDAPServer.jwt_claims(ss_jwt).except(['iss', 'sub', 'exp', 'iat', 'jti']))
-
-        response.body = response_body.to_json
-        response.headers['Cache-Control'] = 'no-store'
-        response.headers['Pragma'] = 'no-cache'
-        response.headers['Access-Control-Allow-Origin'] = '*'
-        response.content_type = 'application/json'
-        response.status = 201
+        make_udap_registration_response
       end
 
       def update_result
@@ -38,19 +25,6 @@ module UDAPSecurityTestKit
 
       def tags
         [REGISTRATION_TAG, UDAP_TAG]
-      end
-
-      private
-
-      def client_uri_from_registration_payload(reg_body)
-        software_statement_jwt = request_software_statement_jwt(reg_body)
-        return unless software_statement_jwt.present?
-
-        MockUDAPServer.jwt_claims(software_statement_jwt)&.dig('iss')
-      end
-
-      def request_software_statement_jwt(reg_body)
-        reg_body&.dig('software_statement')
       end
     end
   end
